+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.3.0
 
-* Add .toArray, .size, .isEmpty, .grep, .zip, .at, .sliceLength, .first, .last, .take, .drop, .map, .each, .filter
+* Lots of new methods
 
 ## v0.2.0